requirements.txt
Django==1.8
djangorestframework==3.1.1
mysql-python==1.2.5
